src/requirements.txt
pycups
python-opencv
Pillow